+++ HISTORY.md
@@ -10,6 +10,7 @@
 * Removed fragment metadata caching.
 * Removed array schema caching.
 * The tile MBR in the in-memory fragment metadata are organized into an R-Tree, speeding up tile overlap operations during subarray reads.
+* Improved encryption key validation process when opening already open arrays. Fixes issue with indefinite growing of the URI to encryption key mapping in `StorageManager` (the mapping is no longer needed). 
 
 ## Bug fixes